+++ cli/src/reporter.rs
@@ -132,7 +132,7 @@ pub fn check_files(
     paths: &[String],
     is_stdin: bool,
     stdin_path: Option<String>,
-    excluded_rules: &[String],
+    excluded_rules: &[RuleViolationKind],
 ) -> Result<Vec<ViolationContent>, CheckFilesError> {
     let mut output_violations = vec![];
 
@@ -587,8 +587,8 @@ mod test_reporter {
 ALTER TABLE "core_foo" ADD COLUMN "bar" integer NOT NULL;
 SELECT 1;
 "#;
-        let violations =
-            check_sql(sql, &["prefer-robust-stmts".into()]).expect("valid sql should parse");
+        let violations = check_sql(sql, &[RuleViolationKind::PreferRobustStmts])
+            .expect("valid sql should parse");
 
         let filename = "main.sql";
 
@@ -614,8 +614,8 @@ SELECT 1;
 ALTER TABLE "core_foo" ADD COLUMN "bar" integer NOT NULL;
 SELECT 1;
 "#;
-        let violations =
-            check_sql(sql, &["prefer-robust-stmts".into()]).expect("valid sql should parse");
+        let violations = check_sql(sql, &[RuleViolationKind::PreferRobustStmts])
+            .expect("valid sql should parse");
         let filename = "main.sql";
         let mut buff = Vec::new();
 
@@ -651,8 +651,8 @@ SELECT 1;
 ALTER TABLE "core_foo" ADD COLUMN "bar" integer NOT NULL;
 SELECT 1;
 "#;
-        let violations =
-            check_sql(sql, &["prefer-robust-stmts".into()]).expect("valid sql should parse");
+        let violations = check_sql(sql, &[RuleViolationKind::PreferRobustStmts])
+            .expect("valid sql should parse");
         let filename = "main.sql";
         let mut buff = Vec::new();
 
@@ -675,8 +675,8 @@ SELECT 1;
 ALTER TABLE "core_foo" ADD COLUMN "bar" integer NOT NULL;
 SELECT 1;
 "#;
-        let violations =
-            check_sql(sql, &["prefer-robust-stmts".into()]).expect("valid sql should parse");
+        let violations = check_sql(sql, &[RuleViolationKind::PreferRobustStmts])
+            .expect("valid sql should parse");
 
         let filename = "main.sql";
         assert_debug_snapshot!(pretty_violations(violations, sql, filename));
@@ -687,8 +687,8 @@ SELECT 1;
     #[test]
     fn test_trimming_sql_newlines() {
         let sql = r#"ALTER TABLE "core_recipe" ADD COLUMN "foo" integer NOT NULL;"#;
-        let violations =
-            check_sql(sql, &["prefer-robust-stmts".into()]).expect("valid sql should parse");
+        let violations = check_sql(sql, &[RuleViolationKind::PreferRobustStmts])
+            .expect("valid sql should parse");
 
         assert_debug_snapshot!(violations, @r###"
         [